@@ -5,20 +5,24 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "php": "^7.2",
-        "fideloper/proxy": "^4.3",
-        "laravel/framework": "^7.6",
-        "laravel/ui": "^2.0",
-        "laravelcollective/html": "^6.1"
+        "php": "^7.3|^8.0",
+        "fideloper/proxy": "^4.4",
+        "fruitcake/laravel-cors": "^2.0",
+        "guzzlehttp/guzzle": "^7.0.1",
+        "inertiajs/inertia-laravel": "^0.3.5",
+        "laravel/framework": "^8.12",
+        "laravel/jetstream": "^2.1",
+        "laravel/sanctum": "^2.6",
+        "laravel/tinker": "^2.5",
+        "tightenco/ziggy": "^1.0"
     },
     "require-dev": {
-        "barryvdh/laravel-ide-helper": "^2.6",
-        "fzaninotto/faker": "^1.9",
-        "laravel/homestead": "^10.7",
-        "laravel/tinker": "^2.4",
-        "mockery/mockery": "^1.3",
-        "nunomaduro/collision": "^4.2",
-        "phpunit/phpunit": "^9.0"
+        "facade/ignition": "^2.5",
+        "fakerphp/faker": "^1.9.1",
+        "laravel/sail": "^1.0.1",
+        "mockery/mockery": "^1.4.2",
+        "nunomaduro/collision": "^5.0",
+        "phpunit/phpunit": "^9.3.3"
     },
     "autoload": {
         "classmap": [
