@@ -7,8 +7,9 @@ return [
     'confirmed' => ':attribute stimmt nicht überein.',
     'min' => [
         'numeric' => ':attribute muss mindestens :min. sein',
-        'file' => ':attribute muss mindestens :min Kilobyte groß sein.',
+        'file' => ':attribute muss mindestens :mdin Kilobyte groß sein.',
         'string' => ':attribute muss mindestens :min Zeichen lang sein.',
         'array' => ':attribute muss mindestens :min Einträge haben.',
     ],
+    'required' => 'Pflichtfeld',
 ];
